Add extra function to cover all the ways of filtering.
