+++ packages/extension/src/view/devtools/stateProviders/filterManagementStore/index.tsx
@@ -30,7 +30,6 @@ import type { CookieTableData } from '../../cookies.types';
 import getFilters from './utils/getFilters';
 import filterCookies from './utils/filterCookies';
 import { noop } from '../../../../utils/noop';
-import filterCookiesWithRetentionPeriod from './utils/filterCookiesWithRetentionPeriod';
 
 export interface filterManagementStore {
   state: {
@@ -96,14 +95,9 @@ export const Provider = ({ children }: PropsWithChildren) => {
     if (!selectedFrame) {
       return [];
     }
-    const retentionPeriodFilteredCookies = filterCookiesWithRetentionPeriod(
-      frameFilteredCookies,
-      selectedFrameFilters[selectedFrame]?.selectedFilters || {},
-      searchTerm
-    );
     return Object.values(
       filterCookies(
-        retentionPeriodFilteredCookies,
+        frameFilteredCookies,
         selectedFrameFilters[selectedFrame]?.selectedFilters || {},
         searchTerm
       )
