+++ packages/extension/src/worker/service-worker.ts
@@ -34,7 +34,7 @@ import {
 let cookieDB: CookieDatabase | null = null;
 
 // Global promise queue.
-const PROMISE_QUEUE = new PQueue({ concurrency: 1 });
+const PROMISE_QUEUE = new PQueue({ concurrency: 100 });
 
 /**
  * Fires when the browser receives a response from a web server.
